Revert "Fix cargo build for x86 (#658)"
This reverts commit 9adc280f04c651aae0d4afb65fa19c2e1e9cac19.
