+++ bazel/external_crates.bzl
@@ -67,7 +67,6 @@ def external_crates_repository():
             "//rs/rollout-controller:Cargo.toml",
             "//rs/dre-canisters/trustworthy-node-metrics/src/trustworthy-node-metrics:Cargo.toml",
             "//rs/qualifier:Cargo.toml",
-            "//rs/dre-canisters/trustworthy-node-metrics/src/trustworthy-node-metrics-types:Cargo.toml",
         ],
         splicing_config = splicing_config(
             resolver_version = "2",
