--- rs/dre-canisters/trustworthy-node-metrics/src/trustworthy-node-metrics-types/BUILD.bazel
@@ -1,16 +1,0 @@
-load("@crate_index_dre//:defs.bzl", "aliases", "all_crate_deps")
-load("@rules_rust//rust:defs.bzl", "rust_library")
-
-package(default_visibility = ["//visibility:public"])
-
-rust_library(
-    name = "trustworthy_node_metrics_types",
-    srcs = glob(["src/**/*.rs"]),
-    aliases = aliases(),
-    proc_macro_deps = all_crate_deps(
-        proc_macro = True,
-    ),
-    deps = all_crate_deps(
-        normal = True,
-    ),
-)
