--- rs/dre-canisters/trustworthy-node-metrics/src/trustworthy-node-metrics-types/src/lib.rs
@@ -1,1 +1,0 @@
-pub mod types;
